Implement protections for deleting past events
Additional checks have been added to the deleteEvent method in the EventService class, now preventing past events from being deleted. A NoSuchElementException has been introduced to handle cases where an event is not found, returning a "Resource not found" response.

+++ src/main/java/org/regikeskus/events/service/EventService.java
@@ -6,7 +6,9 @@ import org.regikeskus.events.repository.EventRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @RequiredArgsConstructor
@@ -28,12 +30,19 @@ public class EventService {
     @Transactional
     public Event createOrUpdateEvent(Event event) {
         if (event.getEventName() == null || event.getEventDateTime() == null || event.getLocation() == null) {
-            throw new IllegalArgumentException("Event name or date/time or location must not be null");  }
+            throw new IllegalArgumentException("Event name or date/time or location must not be null");
+        }
         return eventRepository.save(event);
     }
 
     @Transactional
     public void deleteEvent(Long id) {
+        Event event = eventRepository.findById(id).orElseThrow(() -> new NoSuchElementException("Event not found with ID: " + id));
+
+        if (event.getEventDateTime().isBefore(LocalDateTime.now())) {
+            throw new IllegalArgumentException("Cannot delete past events.");
+        }
+
         eventRepository.deleteById(id);
     }
 }
